requirements.txt
# Web 框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gunicorn>=21.2.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0

# OCR 引擎
paddleocr>=3.0.0
paddlepaddle>=3.0.0           # CPU版本；paddleocr 3.x 需要 paddle 3.x

# 图像处理
opencv-python-headless>=4.8.0   # 服务端无需GUI，用headless版
Pillow>=10.0.0
numpy>=1.24.0

# LLM 客户端
openai>=1.0.0             # OpenAI GPT-4V
anthropic>=0.25.0         # Claude Vision

# HTTP 客户端
httpx>=0.25.0             # 图片 URL 下载

# 对象存储
cos-python-sdk-v5>=1.9.0  # 腾讯云 COS

# 工具库
python-dotenv>=1.0.0      # 环境变量管理

# 测试
pytest>=8.0.0
pytest-cov>=5.0.0
openpyxl>=3.1.0            # Excel 读写
